scripts/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0
scholarly>=1.7.0
crossref-commons>=0.0.7
arxiv>=1.4.0
bibtexparser>=1.4.0
python-dateutil>=2.8.0
pyyaml>=6.0
urllib3>=1.26.0
selenium>=4.15.0
webdriver-manager>=4.0.0